+++ src/models/accinfo.ts
@@ -2,6 +2,7 @@ import { IProxy } from './proxy';
 
 export interface IAccountInfo {
     buildVersion: string;
+    localServer?: ILocalServerSettings;
     accounts: IAccount[];
 }
 export interface IAccount {
@@ -17,4 +18,9 @@ export interface ICharacterInfo {
     charId: number;
     nextCharId: number;
     maxNumChars: number;
+}
+
+export interface ILocalServerSettings {
+    enabled: boolean;
+    port?: number;
 }